% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlaplace}
\alias{dlaplace}
\title{Density function of the Laplace distribution}
\usage{
dlaplace(x, m, b)
}
\arguments{
\item{x}{vector}

\item{m}{location}

\item{b}{scale}
}
\description{
Density function of the Laplace distribution
}
