% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plaplace}
\alias{plaplace}
\title{Distribution function of the Laplace distribution}
\usage{
plaplace(x, m, b)
}
\arguments{
\item{x}{quantile}

\item{m}{location}

\item{b}{scale}
}
\description{
Distribution function of the Laplace distribution
}
