% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Sample from the Dirichlet distribution}
\usage{
rdirichlet(n, a)
}
\arguments{
\item{n}{number ofsample}

\item{a}{vector of weights}
}
\description{
Sample from the Dirichlet distribution
}
