% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdiscrete}
\alias{rdiscrete}
\title{Sample from the discrete distribution}
\usage{
rdiscrete(n, p, as_indicator = FALSE)
}
\arguments{
\item{n}{number of samples}

\item{p}{vector of class probabilities, must sum to 1}

\item{as_indicator}{false: returns class number, true: returns binary vector}
}
\description{
Sample from the discrete distribution
}
