% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgem}
\alias{rgem}
\title{Sample from the GEM distribution}
\usage{
rgem(n, k, b)
}
\arguments{
\item{n}{number of samples}

\item{k}{vector size}

\item{b}{shape}
}
\description{
Sample from the GEM distribution
}
