% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlaplace}
\alias{rlaplace}
\title{Sample from the Laplace distribution}
\usage{
rlaplace(n, m, b)
}
\arguments{
\item{n}{number of samples}

\item{m}{location}

\item{b}{scale}
}
\description{
Sample from the Laplace distribution
}
